// app/dashboard/karbooms/third-party-insurance-list/_schemas/third-party-insurance-form-schema.ts
import { Dayjs } from "dayjs";
import { z } from "zod";

const ThirdPartyInsuranceFormSchema = z.object({
  insurance_company_id: z.number("انتخاب شرکت بیمه الزامی است"),
  insurance_number: z.string("شماره بیمه الزامی است"),
  insurance_code: z.string("کد یکتای بیمه الزامی است"),
  started_at: z.custom<Dayjs>(),
  ended_at: z.custom<Dayjs>(),
  description: z.string().nullish(),
});

export default ThirdPartyInsuranceFormSchema;

export type ThirdPartyInsuranceFormType = z.infer<
  typeof ThirdPartyInsuranceFormSchema
>;